+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add `Container#removeComponents()` to remove all child components of a container
 - Display multiple subtitle cues in parallel in `SubtitleOverlay`
 - Add `getText()` method, `isEmpty()` method, and `onTextChanged` event to `Label`
+- Add `TitleBarConfig#keepHiddenWithoutMetadata` to keep `TitleBar` hidden if metadata labels are empty
 
 ### Fixed
 - Clear `SubtitleOverlay` when playback is finished